+++ src/react/views/modal/DatePickerIOS.js
@@ -21,7 +21,9 @@ class Picker extends React.Component {
 
   render () {
     return <Base visible onRequestClose={this.props.onCancel}>
-      <DatePickerIOS mode='date' date={this.state.date} onDateChange={date => this.setState({date})}/>
+      <DatePickerIOS
+        minimumDate={new Date()}
+        mode='date' date={this.state.date} onDateChange={date => this.setState({date})}/>
       <View style={modal.buttonContainer}>
         <View style={modal.button}>
           <FancyTextButton